Move StyledText out of boot_animation.py and update documentation
